+++ src/components/MydashboardPage/InvitedDashboard/InvitedDashboard.jsx
@@ -1,8 +1,9 @@
-import { useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import Image from 'next/image';
 import { NoMailIcon, SearchIcon } from '@/../public/images';
 import useIntersectionObserver from '@/hooks/useIntersectionObserver';
 import CtaDefault from '../../common/Buttons/CtaDefault/CtaDefault';
+import useDebounce from '@/hooks/useDebounce';
 
 // 송상훈 TODO :invitations가 없을때 맨아래 div가 뜨는데 그럼 로딩때는..?
 export default function InvitedDashboard({
@@ -13,10 +14,22 @@ export default function InvitedDashboard({
 }) {
   const observerRef = useRef(null);
   const scrollContainerRef = useRef(null);
-  const [inputValue, setInputValue] = useState(''); // 입력 값 임시저장
-  const [searchTerm, setSearchTerm] = useState(''); // 확정된 입력 값
+  const [inputValue, setInputValue] = useState('');
+  const debouncedSearchTerm = useDebounce(inputValue, 700);
 
   const isEmpty = !invitations || invitations.length === 0;
+
+  const handleInputChange = (event) => {
+    setInputValue(event.target.value);
+  };
+
+  useEffect(() => {
+    if (debouncedSearchTerm === '') {
+      updateTitle(inputValue);
+    } else {
+      updateTitle(inputValue);
+    }
+  }, [debouncedSearchTerm]);
 
   useIntersectionObserver(observerRef, scrollContainerRef, () => {
     if (!loading) {
@@ -24,54 +37,33 @@ export default function InvitedDashboard({
     }
   });
 
-  const handleInputChange = (event) => {
-    setInputValue(event.target.value); // 입력 값 임시 저장
-  };
-
-  const handleKeyDown = (event) => {
-    if (event.key === 'Enter') {
-      setSearchTerm(inputValue);
-      updateTitle(searchTerm);
-    }
-  };
-
   return (
     <section className="bg-white max-w-[1022px] h-auto rounded-[8px] shadow-sm py-[24px] px-[16px]">
       <h2 className="text-xl font-bold mb-5 ">초대받은 대시보드</h2>
 
-      {!isEmpty && (
-        <>
-          <div className="flex items-center mb-[24px]">
-            <div className="relative w-full ">
-              <input
-                className="w-full rounded-[6px] border-2 h-[40px] pl-10"
-                type="text"
-                placeholder="Search Enter"
-                value={inputValue}
-                onChange={handleInputChange}
-                onKeyDown={handleKeyDown}
-              />
-              <div className="absolute left-3 top-1/2 transform -translate-y-1/2">
-                <Image
-                  src={SearchIcon}
-                  width={24}
-                  height={24}
-                  alt="돋보기아이콘"
-                />
-              </div>
-            </div>
+      <div className="flex items-center mb-[24px]">
+        <div className="relative w-full ">
+          <input
+            className="w-full rounded-[6px] border-2 h-[40px] pl-10"
+            type="text"
+            placeholder="Search"
+            value={inputValue}
+            onChange={handleInputChange}
+          />
+          <div className="absolute left-3 top-1/2 transform -translate-y-1/2">
+            <Image src={SearchIcon} width={24} height={24} alt="돋보기아이콘" />
           </div>
+        </div>
+      </div>
 
-          <div className="hidden md:text-gray_9FA6B2 md:grid md:grid-cols-3 ">
-            <p>이름</p>
-            <p>초대자</p>
-            <p>수락 여부</p>
-          </div>
-        </>
-      )}
+      <div className="hidden md:text-gray_9FA6B2 md:grid md:grid-cols-3 ">
+        <p>이름</p>
+        <p>초대자</p>
+        <p>수락 여부</p>
+      </div>
 
       {isEmpty ? (
-        <div className="h-[300px] flex flex-col justify-center items-center">
+        <div className="h-[400px] flex flex-col justify-center items-center">
           <Image
             src={NoMailIcon}
             width={100}
